refactor unit B to A
