+++ server.js
@@ -1,10 +1,6 @@
 'use strict';
-
 const Glue = require('glue');
-const mongoose = require('mongoose');
-
 const manifest = require('./config/manifest');
-const db = require('./config/config');
 
 if (!process.env.PRODUCTION) {
   manifest.registrations.push({
@@ -23,9 +19,3 @@ Glue.compose(manifest, { relativeTo: __dirname }, (err, server) => {
     console.log('✅  Server is listening on ' + server.info.uri.toLowerCase());
   });
 });
-
- // cek koneksi database 
-//  mongoose.connect(db.database.uri, {useNewUrlParser: true, useUnifiedTopology: true});
-//  mongoose.connection.on('error', console.error.bind(console, 'Connection error.'));
-//  mongoose.connection.once('open', function() {console.log("✅  Connected to database. success!");
-//  });
